requirements.txt
asgiref==3.3.4
asn1crypto==0.24.0
blinker==1.4
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
click==6.7
colorama==0.3.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.5
django-allauth==0.45.0
django-cors-headers==3.7.0
django-heroku==0.3.1
django-storages
datetime
idna
feedparser
httplib2
django-filter==2.4.0
django-multiselectfield==0.1.12
django-rest-auth==0.9.5
django-restframework==0.0.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt
djoser==2.1.0
lazr.restfulclient
lazr.uri
gunicorn
Jinja2
lxml==4.6.3
MarkupSafe
oauth==1.0.1
oauthlib==3.1.1
Pillow==8.3.1
PyJWT
pytz
python3-openid
requests
requests-oauthlib
requests-unixsocket
SecretStorage
six
social-auth-app-django
social-auth-core
sqlparse
soupsieve
uritemplate
urllib3
whitenoise
python-dotenv==0.15.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
